# merchant/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is merchant golang server.",
        "title": "Swagger Merchant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api",
    "paths": {
...
        "/variants": {
            "post": {
                "description": "Create a new variant with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variants"
                ],
                "summary": "Create a new variant",
                "parameters": [
                    {
                        "description": "Variant details",
                        "name": "variant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO"
                        }
                    }
                }
            }
        },
        "/variants/product-id/{id}": {
            "get": {
                "description": "Get a variant by its product ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variants"
                ],
                "summary": "Get a variant by product ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO"
                        }
                    }
                }
            }
        },
        "/variants/variant-name/{variantName}": {
            "get": {
                "description": "Get a variant by its variant name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variants"
                ],
                "summary": "Get a variant by variant name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variant Name",
                        "name": "variantName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO"
                        }
                    }
                }
            }
        },
        "/variants/{id}": {
            "get": {
                "description": "Get a variant by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variants"
                ],
                "summary": "Get a variant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a variant by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Variants"
                ],
                "summary": "Delete a variant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_VuThanhThien_golang-gorm-postgres_merchant_internal_models_dto.VariantDTO": {
            "type": "object",
            "required": [
                "price",
                "product_id",
                "variant_name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "variant_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}